feat: integrate supabase postgres database
