+++ package.json
@@ -31,6 +31,7 @@
     "cors": "^2.8.5",
     "dotenv": "^17.2.2",
     "express": "^5.1.0",
+    "pg": "^8.12.0",
     "zod": "^4.1.9"
   },
   "devDependencies": {
@@ -39,6 +40,7 @@
     "@types/express": "^5.0.3",
     "@types/jest": "^29.5.12",
     "@types/node": "^24.5.2",
+    "@types/pg": "^8.15.5",
     "concurrently": "^9.2.1",
     "jest": "^29.7.0",
     "supertest": "^7.1.1",
