+++ src/index.ts
@@ -2,8 +2,8 @@ import { createApp } from './app';
 import { config } from './config';
 import { initDB } from './db';
 
-function bootstrap(): void {
-  initDB();
+async function bootstrap(): Promise<void> {
+  await initDB();
   const app = createApp();
 
   app.listen(config.port, () => {
@@ -12,4 +12,8 @@ function bootstrap(): void {
   });
 }
 
-bootstrap();
+void bootstrap().catch((error) => {
+  // eslint-disable-next-line no-console
+  console.error('Failed to start application', error);
+  process.exit(1);
+});
