+++ src/routes/projects.ts
@@ -8,16 +8,16 @@ const createProjectSchema = z.object({
   name: z.string().min(2).max(120),
 });
 
-router.get('/', (req, res) => {
+router.get('/', async (req, res) => {
   const search = typeof req.query.search === 'string' ? req.query.search : undefined;
   const limitParam = typeof req.query.limit === 'string' ? Number(req.query.limit) : undefined;
   const limit = Number.isFinite(limitParam) && limitParam ? Math.min(Math.max(limitParam, 1), 100) : 50;
-  const projects = listProjects(search, limit);
+  const projects = await listProjects(search, limit);
 
   res.json({ projects });
 });
 
-router.post('/', (req, res) => {
+router.post('/', async (req, res) => {
   const parseResult = createProjectSchema.safeParse(req.body);
   if (!parseResult.success) {
     res.status(400).json({ error: 'Invalid project payload', details: parseResult.error.flatten() });
@@ -30,7 +30,7 @@ router.post('/', (req, res) => {
     return;
   }
 
-  const project = createProject(parseResult.data.name, user.id);
+  const project = await createProject(parseResult.data.name, user.id);
   res.status(201).json({ project });
 });
